Revert "luminous: pybind/mgr/mgr_module: make rados handle available to all modules"
Reverts: https://github.com/ceph/ceph/pull/23235
Reverts: 45c7170eb68571a3d690460f99392c6753994799
